@@ -34,7 +34,12 @@
     },
     "config": {
         "vendor-dir": ".build/vendor",
-        "bin-dir": ".build/bin"
+        "bin-dir": ".build/bin",
+        "allow-plugins": {
+            "typo3/cms-composer-installers": true,
+            "typo3/class-alias-loader": true,
+            "sbuerk/typo3-cmscomposerinstallers-testingframework-bridge": true
+        }
     },
     "extra": {
         "typo3/cms": {
@@ -44,21 +49,23 @@
         }
     },
     "require": {
-        "php": "^7.2 || ^8.0",
+        "php": "^8.1",
         "ext-pdo": "*",
-        "typo3/cms-core": "^10.4 || ^11.5",
-        "typo3/cms-extbase": "^10.4 || ^11.5",
-        "typo3/cms-fluid": "^10.4 || ^11.5",
-        "extcode/cart": "^8.0"
+        "extcode/cart": "dev-main",
+        "typo3/cms-core": "^12.4",
+        "typo3/cms-extbase": "^12.4",
+        "typo3/cms-fluid": "^12.4"
     },
     "require-dev": {
-        "typo3/testing-framework": "^6.0",
-        "typo3/cms-dashboard": "^10.4 || ^11.5",
-        "friendsofphp/php-cs-fixer": "^2.14",
-        "helmich/typo3-typoscript-lint": "^2.0",
-        "overtrue/phplint": "^1.1",
-        "rector/rector": "^0.11",
-        "phpstan/phpstan": "^0.12.99"
+        "friendsofphp/php-cs-fixer": "^3.16",
+        "helmich/typo3-typoscript-lint": "^3.1",
+        "overtrue/phplint": "^5.5",
+        "phpstan/phpstan": "^1.10",
+        "sbuerk/typo3-cmscomposerinstallers-testingframework-bridge": "^0.0.1",
+        "ssch/typo3-rector": "^1.2",
+        "typo3/cms-dashboard": "^12.4",
+        "typo3/cms-reactions": "^12.4",
+        "typo3/testing-framework": "^8.0"
     },
     "scripts": {
         "test:cgl": [
@@ -77,7 +84,7 @@
             "typo3DatabaseDriver=\"pdo_sqlite\" .build/bin/phpunit -c Build/FunctionalTests.xml"
         ],
         "test:phpstan:analyse": [
-            ".build/bin/phpstan analyse -c Build/phpstan.neon"
+            ".build/bin/phpstan analyse -c Build/phpstan.neon --memory-limit 256M"
         ],
         "test:rector:process": [
             ".build/bin/rector process *"
@@ -99,10 +106,12 @@
             "@test:cgl",
             "@test:typoscript:lint",
             "@test:php"
-        ],
-        "post-autoload-dump": [
-            "mkdir -p .build/public/typo3conf/ext/",
-            "[ -L .build/public/typo3conf/ext/cart_products ] || ln -snvf ../../../../. .build/public/typo3conf/ext/cart_products"
         ]
+    },
+    "repositories": {
+        "extcode/cart": {
+            "type": "vcs",
+            "url": "ssh://git@code.extco.de:22722/cart/cart.git"
+        }
     }
 }
