@@ -9,27 +9,23 @@ namespace Extcode\CartProducts\Domain\Model\Product;
  * LICENSE file that was distributed with this source code.
  */
 
-use TYPO3\CMS\Extbase\Domain\Model\FrontendUserGroup;
+use Extcode\Cart\Domain\Model\FrontendUserGroup;
+use TYPO3\CMS\Extbase\Annotation\Validate;
 use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
 
 class SpecialPrice extends AbstractEntity
 {
     /**
-     * @var string
-     * @TYPO3\CMS\Extbase\Annotation\Validate("NotEmpty")
+     * @Validate("NotEmpty")
      */
-    protected $title = '';
+    protected string $title = '';
 
     /**
-     * @var float
-     * @TYPO3\CMS\Extbase\Annotation\Validate("NotEmpty")
+     * @Validate("NotEmpty")
      */
-    protected $price = 0.0;
+    protected float $price = 0.0;
 
-    /**
-     * @var FrontendUserGroup
-     */
-    protected $frontendUserGroup;
+    protected ?FrontendUserGroup $frontendUserGroup = null;
 
     public function getTitle(): string
     {
