@@ -1,5 +1,7 @@
 <?php
+
 declare(strict_types=1);
+
 namespace Extcode\CartProducts\EventListener;
 
 /*
@@ -9,82 +11,86 @@ namespace Extcode\CartProducts\EventListener;
  * LICENSE file that was distributed with this source code.
  */
 
+use Extcode\Cart\Domain\Model\FrontendUser;
+use Extcode\Cart\Domain\Repository\FrontendUserRepository;
 use Extcode\Cart\Event\RetrieveProductsFromRequestEvent;
 use Extcode\CartProducts\Domain\Repository\Product\ProductRepository;
-use Extcode\CartProducts\Utility\ProductUtility;
-use TYPO3\CMS\Core\Messaging\AbstractMessage;
-use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
+use Psr\EventDispatcher\EventDispatcherInterface;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
 
 class RetrieveProductsFromRequest
 {
-    /**
-     * @var ProductRepository
-     */
-    protected $productRepository;
+    private EventDispatcherInterface $eventDispatcher;
 
-    /**
-     * @var ProductUtility
-     */
-    protected $productUtility;
+    protected ProductRepository $productRepository;
 
     public function __construct(
-        ProductRepository $productRepository,
-        ProductUtility $productUtility
+        EventDispatcherInterface $eventDispatcher,
+        ProductRepository $productRepository
     ) {
+        $this->eventDispatcher = $eventDispatcher;
         $this->productRepository = $productRepository;
-        $this->productUtility = $productUtility;
     }
 
     public function __invoke(RetrieveProductsFromRequestEvent $event): void
     {
         $request = $event->getRequest();
-        $cart = $event->getCart();
+        $taxClasses = $event->getCart()->getTaxClasses();
         $requestArguments = $request->getArguments();
-        $taxClasses = $cart->getTaxClasses();
 
         if ($requestArguments['productType'] !== 'CartProducts') {
             return;
         }
 
-        $errors = $this->checkRequestArguments($requestArguments);
+        $frontendUserGroupIds = $this->getFrontendUserGroupIds();
 
+        $createEvent = new \Extcode\CartProducts\Event\RetrieveProductsFromRequestEvent($request, $taxClasses);
+        $createEvent->setFrontendUserGroupIds($frontendUserGroupIds);
+        $this->eventDispatcher->dispatch($createEvent);
+
+        $errors = $createEvent->getErrors();
         if (!empty($errors)) {
             $event->setErrors($errors);
+            $event->setPropagationStopped(true);
             return;
         }
 
-        $this->productUtility->setTaxClasses($taxClasses);
-
-        $event->addProduct(
-            $this->productUtility->getProductFromRequest(
-                $request,
-                $cart->getTaxClasses()
-            )
-        );
+        $event->addProduct($createEvent->getCartProduct());
     }
 
-    protected function checkRequestArguments(array $requestArguments): array
+    /**
+     * Get Frontend User Group
+     */
+    protected function getFrontendUserGroupIds(): array
     {
-        if (!(int)$requestArguments['product']) {
-            return [
-                'messageBody' => LocalizationUtility::translate(
-                    'tx_cart.error.parameter.no_product',
-                    'cart_products'
-                ),
-                AbstractMessage::ERROR
-            ];
+        $feGroupIds = [];
+
+        $user = $GLOBALS['TSFE']->fe_user->user;
+        if (!$user || !(int)$user['uid']) {
+            return $feGroupIds;
         }
 
-        if ((int)$requestArguments['quantity'] < 0) {
-            return [
-                'messageBody' => LocalizationUtility::translate(
-                    'tx_cart.error.invalid_quantity',
-                    'cart_products'
-                ),
-                'severity' => AbstractMessage::WARNING
-            ];
+        $feGroups = $this->getFeGroups((int)$user['uid']);
+        if ($feGroups) {
+            foreach ($feGroups as $feGroup) {
+                $feGroupIds[] = $feGroup->getUid();
+            }
         }
 
-        return [];
+        return $feGroupIds;
+    }
+
+    protected function getFeGroups(int $uid): ?ObjectStorage
+    {
+        $frontendUserRepository = GeneralUtility::makeInstance(
+            FrontendUserRepository::class
+        );
+        $feUser = $frontendUserRepository->findByUid((int)$uid);
+        if (!$feUser instanceof FrontendUser) {
+            return null;
+        }
+
+        return $feUser->getUsergroup();
     }
 }
