@@ -1,5 +1,7 @@
 <?php
+
 declare(strict_types=1);
+
 namespace Extcode\CartProducts\EventListener\Order\Stock;
 
 /*
@@ -28,11 +30,7 @@ class HandleStock
 
                 $product = $productQueryBuilder
                     ->select('uid', 'handle_stock', 'handle_stock_in_variants')
-                    ->from('tx_cartproducts_domain_model_product_product')
-                    ->where(
-                        $productQueryBuilder->expr()->eq('uid', $productQueryBuilder->createNamedParameter($cartProduct->getProductId(), \PDO::PARAM_INT))
-                    )
-                    ->execute()->fetch();
+                    ->from('tx_cartproducts_domain_model_product_product')->where($productQueryBuilder->expr()->eq('uid', $productQueryBuilder->createNamedParameter($cartProduct->getProductId(), \PDO::PARAM_INT)))->executeQuery()->fetch();
 
                 if ($product['handle_stock']) {
                     if ($product['handle_stock_in_variants']) {
@@ -53,11 +51,7 @@ class HandleStock
 
         $product = $productQueryBuilder
             ->select('stock')
-            ->from('tx_cartproducts_domain_model_product_product')
-            ->where(
-                $productQueryBuilder->expr()->eq('uid', $productQueryBuilder->createNamedParameter($cartProduct->getProductId(), \PDO::PARAM_INT))
-            )
-            ->execute()->fetch();
+            ->from('tx_cartproducts_domain_model_product_product')->where($productQueryBuilder->expr()->eq('uid', $productQueryBuilder->createNamedParameter($cartProduct->getProductId(), \PDO::PARAM_INT)))->executeQuery()->fetch();
 
         $productQueryBuilder
             ->update('tx_cartproducts_domain_model_product_product')
@@ -66,9 +60,7 @@ class HandleStock
             )
             ->orWhere(
                 $productQueryBuilder->expr()->eq('l10n_parent', $productQueryBuilder->createNamedParameter($cartProduct->getProductId(), \PDO::PARAM_INT))
-            )
-            ->set('stock', $product['stock'] - $cartProduct->getQuantity())
-            ->execute();
+            )->set('stock', $product['stock'] - $cartProduct->getQuantity())->executeStatement();
     }
 
     protected function handleStockInBeVariant(CartProduct $cartProduct): void
@@ -80,11 +72,7 @@ class HandleStock
             $beVariantQueryBuilder = $beVariantConnection->createQueryBuilder();
             $beVariant = $beVariantQueryBuilder
                 ->select('stock')
-                ->from('tx_cartproducts_domain_model_product_bevariant')
-                ->where(
-                    $beVariantQueryBuilder->expr()->eq('uid', $beVariantQueryBuilder->createNamedParameter($cartBeVariant->getId(), \PDO::PARAM_INT))
-                )
-                ->execute()->fetch();
+                ->from('tx_cartproducts_domain_model_product_bevariant')->where($beVariantQueryBuilder->expr()->eq('uid', $beVariantQueryBuilder->createNamedParameter($cartBeVariant->getId(), \PDO::PARAM_INT)))->executeQuery()->fetch();
 
             $beVariantQueryBuilder
                 ->update('tx_cartproducts_domain_model_product_bevariant')
@@ -93,9 +81,7 @@ class HandleStock
                 )
                 ->orWhere(
                     $beVariantQueryBuilder->expr()->eq('l10n_parent', $beVariantQueryBuilder->createNamedParameter($cartBeVariant->getId(), \PDO::PARAM_INT))
-                )
-                ->set('stock', $beVariant['stock'] - $cartBeVariant->getQuantity())
-                ->execute();
+                )->set('stock', $beVariant['stock'] - $cartBeVariant->getQuantity())->executeStatement();
         }
     }
 }
