@@ -31,7 +31,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getProductInitiallyReturnsNull()
     {
-        $this->assertNull(
+        self::assertNull(
             $this->beVariant->getProduct()
         );
     }
@@ -44,7 +44,7 @@ class BeVariantTest extends UnitTestCase
         $product = new Product();
         $this->beVariant->setProduct($product);
 
-        $this->assertSame(
+        self::assertSame(
             $product,
             $this->beVariant->getProduct()
         );
@@ -55,7 +55,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getBeVariantAttributeOption1InitiallyIsNull()
     {
-        $this->assertNull(
+        self::assertNull(
             $this->beVariant->getBeVariantAttributeOption1()
         );
     }
@@ -69,7 +69,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->setBeVariantAttributeOption1($beVariantAttributeOption);
 
-        $this->assertSame(
+        self::assertSame(
             $beVariantAttributeOption,
             $this->beVariant->getBeVariantAttributeOption1()
         );
@@ -80,7 +80,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getBeVariantAttributeOption2InitiallyIsNull()
     {
-        $this->assertNull(
+        self::assertNull(
             $this->beVariant->getBeVariantAttributeOption2()
         );
     }
@@ -94,7 +94,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->setBeVariantAttributeOption2($beVariantAttributeOption);
 
-        $this->assertSame(
+        self::assertSame(
             $beVariantAttributeOption,
             $this->beVariant->getBeVariantAttributeOption2()
         );
@@ -105,7 +105,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getBeVariantAttributeOption3InitiallyIsNull()
     {
-        $this->assertNull(
+        self::assertNull(
             $this->beVariant->getBeVariantAttributeOption3()
         );
     }
@@ -119,7 +119,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->setBeVariantAttributeOption3($beVariantAttributeOption);
 
-        $this->assertSame(
+        self::assertSame(
             $beVariantAttributeOption,
             $this->beVariant->getBeVariantAttributeOption3()
         );
@@ -132,7 +132,7 @@ class BeVariantTest extends UnitTestCase
     {
         $price = 0.0;
 
-        $this->assertSame(
+        self::assertSame(
             $price,
             $this->beVariant->getPrice()
         );
@@ -147,7 +147,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->setPrice($price);
 
-        $this->assertSame(
+        self::assertSame(
             $price,
             $this->beVariant->getPrice()
         );
@@ -158,7 +158,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getSpecialPricesInitiallyReturnsEmptyObjectStorage()
     {
-        $this->assertInstanceOf(
+        self::assertInstanceOf(
             ObjectStorage::class,
             $this->beVariant->getSpecialPrices()
         );
@@ -173,7 +173,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->addSpecialPrice($specialPrice);
 
-        $this->assertSame(
+        self::assertSame(
             $specialPrice,
             $this->beVariant->getBestSpecialPrice()
         );
@@ -188,7 +188,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->addSpecialPrice($specialPrice);
 
-        $this->assertSame(
+        self::assertSame(
             $specialPrice,
             $this->beVariant->getBestSpecialPrice()
         );
@@ -199,27 +199,27 @@ class BeVariantTest extends UnitTestCase
      *
      * @return array
      */
-    public function bestSpecialPriceProvider()
+    public static function bestSpecialPriceProvider()
     {
         return [
-            [0,  50.0, 100.0,  150.0, 1],
-            [0, 100.0,  50.0,  150.0, 2],
-            [0, 100.0, 150.0,   50.0, 3],
-            [1,  50.0, 100.0,  150.0, 1],
-            [1, 100.0,  50.0,  150.0, 2],
-            [1, 100.0, 150.0,   50.0, 3],
-            [2, 150.0,  50.0,  100.0, 1],
-            [2,  50.0, 150.0,  100.0, 2],
-            [2, 100.0,  50.0,  150.0, 3],
-            [3, 150.0,  50.0,  100.0, 1],
-            [3,  50.0, 150.0,  100.0, 2],
-            [3, 100.0,  50.0,  150.0, 3],
-            [4,  50.0, 100.0,  150.0, 1],
-            [4, 100.0,  50.0,  150.0, 2],
-            [4, 100.0, 150.0,   50.0, 3],
-            [5,  50.0, 100.0,  150.0, 1],
-            [5, 100.0,  50.0,  150.0, 2],
-            [5, 100.0, 150.0,   50.0, 3],
+            [0, 50.0, 100.0, 150.0, 1],
+            [0, 100.0, 50.0, 150.0, 2],
+            [0, 100.0, 150.0, 50.0, 3],
+            [1, 50.0, 100.0, 150.0, 1],
+            [1, 100.0, 50.0, 150.0, 2],
+            [1, 100.0, 150.0, 50.0, 3],
+            [2, 150.0, 50.0, 100.0, 1],
+            [2, 50.0, 150.0, 100.0, 2],
+            [2, 100.0, 50.0, 150.0, 3],
+            [3, 150.0, 50.0, 100.0, 1],
+            [3, 50.0, 150.0, 100.0, 2],
+            [3, 100.0, 50.0, 150.0, 3],
+            [4, 50.0, 100.0, 150.0, 1],
+            [4, 100.0, 50.0, 150.0, 2],
+            [4, 100.0, 150.0, 50.0, 3],
+            [5, 50.0, 100.0, 150.0, 1],
+            [5, 100.0, 50.0, 150.0, 2],
+            [5, 100.0, 150.0, 50.0, 3],
         ];
     }
 
@@ -253,7 +253,7 @@ class BeVariantTest extends UnitTestCase
             3 => $specialPrice3,
         ];
 
-        $this->assertSame(
+        self::assertSame(
             $specialPrices[$expectedBestSpecialPrice],
             $this->beVariant->getBestSpecialPrice()
         );
@@ -268,7 +268,7 @@ class BeVariantTest extends UnitTestCase
 
         $this->beVariant->setPrice($price);
 
-        $this->assertSame(
+        self::assertSame(
             $price,
             $this->beVariant->getBestPrice()
         );
@@ -279,20 +279,20 @@ class BeVariantTest extends UnitTestCase
      *
      * @return array
      */
-    public function bestPriceProvider()
+    public static function bestPriceProvider()
     {
         return [
-            [0, 100.0,  50.0,  50.0],
+            [0, 100.0, 50.0, 50.0],
             [0, 100.0, 150.0, 100.0],
-            [1, 100.0,  50.0,  50.0],
+            [1, 100.0, 50.0, 50.0],
             [1, 100.0, 150.0, 100.0],
-            [2, 100.0,  50.0, 100.0],
+            [2, 100.0, 50.0, 100.0],
             [2, 100.0, 150.0, 150.0],
-            [3, 100.0,  50.0, 100.0],
+            [3, 100.0, 50.0, 100.0],
             [3, 100.0, 150.0, 150.0],
-            [4, 100.0,  50.0,  50.0],
+            [4, 100.0, 50.0, 50.0],
             [4, 100.0, 150.0, 100.0],
-            [5, 100.0,  50.0,  50.0],
+            [5, 100.0, 50.0, 50.0],
             [5, 100.0, 150.0, 100.0],
         ];
     }
@@ -310,13 +310,13 @@ class BeVariantTest extends UnitTestCase
         $specialPriceObj = $this->createMock(
             SpecialPrice::class
         );
-        $specialPriceObj->expects($this->any())->method('getPrice')->will($this->returnValue($specialPrice));
+        $specialPriceObj->expects(self::any())->method('getPrice')->willReturn($specialPrice);
 
         $this->beVariant->setPrice($price);
         $this->beVariant->setPriceCalcMethod($priceCalcMethod);
         $this->beVariant->addSpecialPrice($specialPriceObj);
 
-        $this->assertSame(
+        self::assertSame(
             $expectedBestPrice,
             $this->beVariant->getBestPrice()
         );
@@ -327,21 +327,21 @@ class BeVariantTest extends UnitTestCase
      *
      * @return array
      */
-    public function bestPriceCalculatedProvider()
+    public static function bestPriceCalculatedProvider()
     {
         return [
             [0, 500.0, 400.0, 350.0, 500.0],
             [0, 500.0, 400.0, 450.0, 500.0],
             [1, 500.0, 400.0, 350.0, 350.0],
             [1, 500.0, 400.0, 450.0, 400.0],
-            [2, 500.0,  20.0,  15.0, 480.0],
-            [2, 500.0,  20.0,  25.0, 475.0],
-            [3, 500.0,  20.0,  15.0, 400.0],
-            [3, 500.0,  20.0,  25.0, 375.0],
-            [4, 500.0,  20.0,  15.0, 515.0],
-            [4, 500.0,  20.0,  25.0, 520.0],
-            [5, 500.0,  20.0,  15.0, 575.0],
-            [5, 500.0,  20.0,  25.0, 600.0],
+            [2, 500.0, 20.0, 15.0, 480.0],
+            [2, 500.0, 20.0, 25.0, 475.0],
+            [3, 500.0, 20.0, 15.0, 400.0],
+            [3, 500.0, 20.0, 25.0, 375.0],
+            [4, 500.0, 20.0, 15.0, 515.0],
+            [4, 500.0, 20.0, 25.0, 520.0],
+            [5, 500.0, 20.0, 15.0, 575.0],
+            [5, 500.0, 20.0, 25.0, 600.0],
         ];
     }
 
@@ -359,7 +359,7 @@ class BeVariantTest extends UnitTestCase
         $specialPriceObj = $this->createMock(
             SpecialPrice::class
         );
-        $specialPriceObj->expects($this->any())->method('getPrice')->will($this->returnValue($specialPrice));
+        $specialPriceObj->expects(self::any())->method('getPrice')->willReturn($specialPrice);
 
         $this->beVariant->setPrice($price);
         $this->beVariant->setPriceCalcMethod($priceCalcMethod);
@@ -369,7 +369,7 @@ class BeVariantTest extends UnitTestCase
         $product->setPrice($productPrice);
         $this->beVariant->setProduct($product);
 
-        $this->assertSame(
+        self::assertSame(
             $expectedBestPrice,
             $this->beVariant->getBestPriceCalculated()
         );
@@ -380,7 +380,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getStockInitiallyReturnsZero()
     {
-        $this->assertSame(
+        self::assertSame(
             0,
             $this->beVariant->getStock()
         );
@@ -394,7 +394,7 @@ class BeVariantTest extends UnitTestCase
         $stock = 10;
         $this->beVariant->setStock($stock);
 
-        $this->assertSame(
+        self::assertSame(
             $stock,
             $this->beVariant->getStock()
         );
@@ -405,7 +405,7 @@ class BeVariantTest extends UnitTestCase
      */
     public function getIsAvailableInitiallyReturnsFalse()
     {
-        $this->assertFalse(
+        self::assertFalse(
             $this->beVariant->getIsAvailable()
         );
     }
@@ -417,7 +417,7 @@ class BeVariantTest extends UnitTestCase
     {
         $this->beVariant->setStock(10);
 
-        $this->assertTrue(
+        self::assertTrue(
             $this->beVariant->getIsAvailable()
         );
     }
