@@ -2,68 +2,40 @@
 
 declare(strict_types=1);
 
+use Rector\Config\RectorConfig;
 use Rector\Core\Configuration\Option;
 use Rector\Core\ValueObject\PhpVersion;
-use Rector\Php55\Rector\String_\StringClassNameToClassConstantRector;
 use Rector\PostRector\Rector\NameImportingPostRector;
 use Ssch\TYPO3Rector\Configuration\Typo3Option;
-use Ssch\TYPO3Rector\FileProcessor\Composer\Rector\ExtensionComposerRector;
-use Ssch\TYPO3Rector\FileProcessor\TypoScript\Visitors\FileIncludeToImportStatementVisitor;
-use Ssch\TYPO3Rector\Rector\General\ConvertTypo3ConfVarsRector;
+use Ssch\TYPO3Rector\Rector\General\ConvertImplicitVariablesToExplicitGlobalsRector;
 use Ssch\TYPO3Rector\Rector\General\ExtEmConfRector;
-use Ssch\TYPO3Rector\Rector\v9\v0\InjectAnnotationRector;
-use Ssch\TYPO3Rector\Set\Typo3SetList;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Ssch\TYPO3Rector\Set\Typo3LevelSetList;
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $parameters = $containerConfigurator->parameters();
+return static function (RectorConfig $rectorConfig): void {
+    // If you want to override the number of spaces for your typoscript files you can define it here, the default value is 4
+    // $parameters = $rectorConfig->parameters();
+    // $parameters->set(Typo3Option::TYPOSCRIPT_INDENT_SIZE, 2);
 
-    $containerConfigurator->import(Typo3SetList::TYPO3_76);
-    $containerConfigurator->import(Typo3SetList::TYPO3_87);
-    $containerConfigurator->import(Typo3SetList::TYPO3_95);
-    $containerConfigurator->import(Typo3SetList::TYPO3_104);
-
-    // In order to have a better analysis from phpstan we teach it here some more things
-    $parameters->set(Option::PHPSTAN_FOR_RECTOR_PATH, Typo3Option::PHPSTAN_FOR_RECTOR_PATH);
-
-    // FQN classes are not imported by default. If you don't do it manually after every Rector run, enable it by:
-    $parameters->set(Option::AUTO_IMPORT_NAMES, true);
-
-    // this will not import root namespace classes, like \DateTime or \Exception
-    $parameters->set(Option::IMPORT_SHORT_CLASSES, false);
-
-    // this will not import classes used in PHP DocBlocks, like in /** @var \Some\Class */
-    $parameters->set(Option::IMPORT_DOC_BLOCKS, false);
+    $rectorConfig->sets([
+        Typo3LevelSetList::UP_TO_TYPO3_12,
+    ]);
 
     // Define your target version which you want to support
-    $parameters->set(Option::PHP_VERSION_FEATURES, PhpVersion::PHP_72);
-
-    // If you have an editorconfig and changed files should keep their format enable it here
-    // $parameters->set(Option::ENABLE_EDITORCONFIG, true);
+    $rectorConfig->phpVersion(PhpVersion::PHP_74);
 
     // If you only want to process one/some TYPO3 extension(s), you can specify its path(s) here.
     // If you use the option --config change __DIR__ to getcwd()
-    // $parameters->set(Option::PATHS, [
+    // $rectorConfig->paths([
     //    __DIR__ . '/packages/acme_demo/',
     // ]);
 
-    // If you set option Option::AUTO_IMPORT_NAMES to true, you should consider excluding some TYPO3 files.
-    // If you use the option --config change __DIR__ to getcwd()
-    $parameters->set(Option::SKIP, [
-        NameImportingPostRector::class => [
-            'ClassAliasMap.php',
-            'ext_emconf.php',
-            'ext_localconf.php',
-            'ext_tables.php',
-            __DIR__ . '/**/Configuration/AjaxRoutes.php',
-            __DIR__ . '/**/Configuration/Backend/AjaxRoutes.php',
-            __DIR__ . '/**/Configuration/Commands.php',
-            __DIR__ . '/**/Configuration/ExpressionLanguage.php',
-            __DIR__ . '/**/Configuration/Extbase/Persistence/Classes.php',
-            __DIR__ . '/**/Configuration/RequestMiddlewares.php',
-            __DIR__ . '/**/Configuration/TCA/*',
-        ],
+    // When you use rector there are rules that require some more actions like creating UpgradeWizards for outdated TCA types.
+    // To fully support you we added some warnings. So watch out for them.
 
+    // If you use importNames(), you should consider excluding some TYPO3 files.
+    $rectorConfig->skip([
+        // @see https://github.com/sabbelasichon/typo3-rector/issues/2536
+        __DIR__ . '/**/Configuration/ExtensionBuilder/*',
         // We skip those directories on purpose as there might be node_modules or similar
         // that include typescript which would result in false positive processing
         __DIR__ . '/**/Resources/**/node_modules/*',
@@ -71,6 +43,18 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         __DIR__ . '/**/Resources/**/BowerComponents/*',
         __DIR__ . '/**/Resources/**/bower_components/*',
         __DIR__ . '/**/Resources/**/build/*',
+        __DIR__ . '/vendor/*',
+        __DIR__ . '/Build/*',
+        __DIR__ . '/public/*',
+        __DIR__ . '/.github/*',
+        __DIR__ . '/.Build/*',
+        NameImportingPostRector::class => [
+            'ext_localconf.php',
+            'ext_tables.php',
+            'ClassAliasMap.php',
+            __DIR__ . '/**/Configuration/*.php',
+            __DIR__ . '/**/Configuration/**/*.php',
+        ],
     ]);
 
     // If you have trouble that rector cannot run because some TYPO3 constants are not defined add an additional constants file
@@ -80,35 +64,30 @@ return static function (ContainerConfigurator $containerConfigurator): void {
     //    __DIR__ . '/typo3.constants.php'
     // ]);
 
-    // get services (needed for register a single rule)
-    $services = $containerConfigurator->services();
-
     // register a single rule
-    // $services->set(InjectAnnotationRector::class);
+    // $rectorConfig->rule(\Ssch\TYPO3Rector\Rector\v9\v0\InjectAnnotationRector::class);
 
     /**
      * Useful rule from RectorPHP itself to transform i.e. GeneralUtility::makeInstance('TYPO3\CMS\Core\Log\LogManager')
      * to GeneralUtility::makeInstance(\TYPO3\CMS\Core\Log\LogManager::class) calls.
      * But be warned, sometimes it produces false positives (edge cases), so watch out
      */
-    // $services->set(StringClassNameToClassConstantRector::class);
+    // $rectorConfig->rule(\Rector\Php55\Rector\String_\StringClassNameToClassConstantRector::class);
 
     // Optional non-php file functionalities:
     // @see https://github.com/sabbelasichon/typo3-rector/blob/main/docs/beyond_php_file_processors.md
 
-    // Adapt your composer.json dependencies to the latest available version for the defined SetList
-    // $containerConfigurator->import(Typo3SetList::COMPOSER_PACKAGES_104_CORE);
-    // $containerConfigurator->import(Typo3SetList::COMPOSER_PACKAGES_104_EXTENSIONS);
-
     // Rewrite your extbase persistence class mapping from typoscript into php according to official docs.
-    // This processor will create a summarized file with all of the typoscript rewrites combined into a single file.
-    // The filename can be passed as argument, "Configuration_Extbase_Persistence_Classes.php" is default.
-    // $services->set(ExtbasePersistenceVisitor::class);
+    // This processor will create a summarized file with all the typoscript rewrites combined into a single file.
+    /* $rectorConfig->ruleWithConfiguration(\Ssch\TYPO3Rector\FileProcessor\TypoScript\Rector\v10\v0\ExtbasePersistenceTypoScriptRector::class, [
+        \Ssch\TYPO3Rector\FileProcessor\TypoScript\Rector\v10\v0\ExtbasePersistenceTypoScriptRector::FILENAME => __DIR__ . '/packages/acme_demo/Configuration/Extbase/Persistence/Classes.php',
+    ]); */
     // Add some general TYPO3 rules
-    $services->set(ConvertTypo3ConfVarsRector::class);
-    $services->set(ExtEmConfRector::class);
-    $services->set(ExtensionComposerRector::class);
+    $rectorConfig->rule(ConvertImplicitVariablesToExplicitGlobalsRector::class);
+    $rectorConfig->ruleWithConfiguration(ExtEmConfRector::class, [
+        ExtEmConfRector::ADDITIONAL_VALUES_TO_BE_REMOVED => [],
+    ]);
 
-    // Do you want to modernize your TypoScript include statements for files and move from <INCLUDE /> to @import use the FileIncludeToImportStatementVisitor
-    // $services->set(FileIncludeToImportStatementVisitor::class);
+    // Modernize your TypoScript include statements for files and move from <INCLUDE /> to @import use the FileIncludeToImportStatementVisitor (introduced with TYPO3 9.0)
+    // $rectorConfig->rule(\Ssch\TYPO3Rector\FileProcessor\TypoScript\Rector\v9\v0\FileIncludeToImportStatementTypoScriptRector::class);
 };
